Add new properties to UserAnimeList and update DB schema
Added new properties to the UserAnimeListDTO and UserAnimeList
classes: EpisodeWatched, StartWatching, FinishWatching, and
DateAdded. Updated OuramDbContextModelSnapshot to reflect these
changes in the UserAnimeLists table. Modified appsettings.json
to change the DefaultConnection string to use localhost instead
of ouram-db. Created a new migration 20240922140059_UserAnimeListNew
to add the new columns to the UserAnimeLists table and generated
the corresponding designer file.